+++ crates/unleash-edge-client-api/src/lib.rs
@@ -12,40 +12,6 @@ pub mod metrics;
 pub mod register;
 pub mod streaming;
 
-fn get_feature_filter(
-    edge_token: &EdgeToken,
-    token_cache: &TokenCache,
-    filter_query: FeatureFilters,
-) -> EdgeResult<(
-    EdgeToken,
-    FeatureFilterSet,
-    unleash_types::client_features::Query,
-)> {
-    info!("Checking {edge_token:?}");
-    let validated_token = token_cache
-        .get(&edge_token.token)
-        .map(|e| e.value().clone())
-        .ok_or(EdgeError::AuthorizationDenied)?;
-
-    let query_filters = filter_query;
-    let query = unleash_types::client_features::Query {
-        tags: None,
-        projects: Some(validated_token.projects.clone()),
-        name_prefix: query_filters.name_prefix.clone(),
-        environment: validated_token.environment.clone(),
-        inline_segment_constraints: Some(false),
-    };
-
-    let filter_set = if let Some(name_prefix) = query_filters.name_prefix {
-        FeatureFilterSet::from(Box::new(name_prefix_filter(name_prefix)))
-    } else {
-        FeatureFilterSet::default()
-    }
-    .with_filter(project_filter(&validated_token));
-
-    Ok((validated_token, filter_set, query))
-}
-
 pub fn router() -> Router<AppState> {
     Router::new()
         .merge(features::router())
