+++ crates/unleash-edge-streaming/Cargo.toml
@@ -12,19 +12,21 @@ axum = { workspace = true }
 dashmap = { workspace = true }
 eventsource-client = { workspace = true }
 futures = { workspace = true }
+futures-util = { workspace = true }
 http = { workspace = true }
 prometheus = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 tokio-stream = { workspace = true }
 tracing = { workspace = true }
-unleash-types = { workspace = true }
 unleash-edge-delta = { path = "../unleash-edge-delta" }
 unleash-edge-feature-cache = { path = "../unleash-edge-feature-cache" }
 unleash-edge-feature-filters = { path = "../unleash-edge-feature-filters" }
+unleash-edge-feature-refresh = { path = "../unleash-edge-feature-refresh" }
 unleash-edge-http-client = { path = "../unleash-edge-http-client" }
 unleash-edge-persistence = { path = "../unleash-edge-persistence" }
 unleash-edge-types = { path = "../unleash-edge-types" }
+unleash-types = { workspace = true }
 unleash-yggdrasil = { workspace = true}
 
 [dev-dependencies]
